pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "olapels",email = "martinsajakaye53@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.124.0,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "google-genai (>=1.53.0,<2.0.0)",
    "pdfplumber (>=0.11.8,<0.12.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
